Version 0.1.008, new music, various optimizations and more.
Altered the map a bit, it's getting closer and closer to a representable
starting zone.
Large Fireball no longer causes a short freeze every time it's fired, I
found what the problem was and I'll continue to add further
optimizations that will make the gameplay even smoother for the next
update.
Added a more recognizable game over situation with its own music, will
work on making it a bit more respectable in the near future.