Version: 0.1.015
Dashing is no longer possible in the air, only when standing/walking on
the ground.
Enemies can no longer harm you while you're in a conversation, be it
reading a sign or talking to someone/something else.
The ConversationBox.java class is no longer static and therefore
supports more than one player.