Version: 0.1.017
Added a pause screen and options/help screen.
The pause screen allows you to pause the game temporarily awaiting your
instruction to resume, it also gives you access to the "Help" message
that describes the basic controls in the game, the Help option will
later on become the options option, allowing you to customize your
gaming experience.
The pause screen further allows you to go back to the menu screen as
well as shut down the game.

Keep in mind that there is still no means of saving your current
progress, that will be added at a later date.