Version: 0.1.019
Changes to the dungeon map, from now on known as "The Mysterious Cave".
The Mysterious Cave:
- Added a couple of enemies.
- Created a basic boss room, will likely undergo further changes.
- Created a very simple boss with some temporary dialogue.
- Actually close the "door" behind you as the conversation starts!
(Wasn't as hard as I first thought it would be.)

The fight itself is simple since it's just another succubus with a
unique name, the main focus here was implementing the means of starting
the fight to begin with! =)