Version: 0.1.39
This is mostly a preparation patch with stuff I want committed while
it's still stable, I intend to try messing around a lot with the code
after this and want to make sure at least this stuff makes it in case I
need to discard everything later.

Added:
- Re-implemented the sounds Fiona used to use and gave them to the
Succubus.

Changed:
- MysteriousDungeon has undergone some changes, probably not final but
it's a little bit better than the previous version.