Version: 0.1.049.5
Oh dear, so many changes, difficult to keep track of them all...

Changes:
- Items should now be saved (but not yet loaded) into the save file.
- All forms of healing should now happen over time rather than instantly
(instant potions are simply way too over powered).
- Altered the conversations with Liadrin and Fiona, as well as the
shrine that activates the fight with Fiona.
- Further fleshed out the ability to make choices in conversations, as
well as included a choice in Liadrin's conversation as well as when you
interact with Fiona's shrine.

Disclaimer:
This version might be a little buggy due to the large amounts of
changes, I will be working through the patch and try to resolve all the
bugs as soon as possible.

Note:
- As the conversations might suggest, I'm starting to get an idea of
what the story will be like (yey!), expect more centralized story from
now on as I now have better ground to stand on.