Version: 0.1.051
Added:
- Added another information piece to the regular hud, right now it's
only used in the tutorial but it can be used later to track quests or
the like.

Changed:
- Vastly improved the tutorial experience by turning the player
invulnerable during its course as well as doing a better job of giving
the player information of what to do.
- Altered the first map a bit as well as removed the old cave.
- All succubi in the first map have 25 % chance to drop a random potion.
- The slug has 100 % chance to drop a healing potion.
- Removed all chests except one in LorasCavern.
- Added one key to LorasCavern.

Known issues:
- Apparently the background causes heavy frame drops in some maps, (was
the case especially in MysteriousDungeon), I'll look into finding some
way of increasing its performance.