Version: 0.1.056
Added:
- New mouse support for the player character, you can now move by
holding down left click and moving the mouse to either the left or right
of your character.
- New mouse support for aiming spells, right now your character will aim
towards wherever your cursor is, this will be changed into an option as
soon as I get a handle on the options menu.
- New mouse support for conversations, conversations now support the
mouse, you can both progress through them by clicking anywhere in the
conversation box, as well as make decisions by hovering over a decision
before clicking.

Note:
- Choosing between playing with the mouse or with the arrow keys is
meant to be an option as soon as I can get it implemented into the
options menu.