Version: 0.1.061
Added:
- New mosue tracking system, it now detects the mouse relative to the
game and not the screen. It is now much more accurate and actually
usable.

Changes:
- It's now possible to move items in the inventory using the mouse, just
drag them from one slot to another. (It will be possible to move them
using the keyboard later.)

Bug fixes:
- Aiming spells with the mouse as well as controlling the player with
the mouse near the edge of the map, is now just as accurate as
everywhere else, no more oddly angled spells.