Version: 0.1.069
So... much... code... =)

Changes:
- Saving and loading should now be much more functional and properly
save and load all units and doodads, including the items they carry and
whether or not they have been used yet. This should even work when
moving between maps, so if you go to FionasSanctum from
MysteriousDungeon and then back again, MysteriousDungeon will "remember"
everything you've done and load it in such a state.

Known issues:
- Items out in the world are not yet saved or loaded, as such the boss
key in MysteriousDungeon will not appear if you load it rather than zone
into it for the first time.