Version: 0.1.073
Changes:
- Explosions now deal damage based on how far away you are from them,
meaning if you're hit dead on, you take close to full damage, but if
you're right at the very edge you might only take about 20-30 % of the
total damage. It might not be perfect yet but it seems to work fairly
well. It applies both to yourself and everyone else.