Version: 0.1.076
Changes:
- Loads of behind the scenes cleanup and optimization to reduce the
sizes of classes, streamlining of things so that they work the same way
as well as moved various event parts out of their respective class into
their respective map to make it easier to find and debug. That way it's
easier to make changes and debug it if something goes wrong. It will
also be easier to reuse it if needed later on.