Version: 0.1.078
Added:
- A new "ability", sprint! You can now make better use of your stamina
as well as make your way through the world faster by holding shift.
Currently it burns a lot of stamina and you will lose the extra speed
the moment you jump (it is not intended to make long jumps, you got
glide and later flight for that).

Changes:
- Removed the extra maps that previously "opened" their respective door
and replaced it with a much better  and smaller solution, simply
replacing the tiles the "door" consists of.
- Water now actually affect the player as well as any other map object
(doodads, units, items), you will move more slowly in water and later on
I will add a breath meter.
- Numerous smaller optimizations and bug fixes.