Version: 0.1.080
Changes:
- I restructured the order things happen a bit when you open a chest,
after unlocking it you will need to press it an additional time to open
it, after that it will now wait for the chest to open fully before
giving you the items. This creates some time between the different
sounds and I feel it makes more sense for you to loot items after the
chest has opened, not as it is opening... ^^
- Made some preparations for more data to be saved to doodads, so that I
can track the changes the player makes to the world (such as opening
cell doors or the like).

Bug fixes:
- The lever that opens Liadrin's door should no longer prevent you from
ending its conversation.
- Chests and doors should now animate properly again.