Version: 0.1.084
Added:
- New art for the exit button (will probably edit this one further later
as it doesn't go well with the rest of the art, but it's better than
nothing).
- New art for the edit key-bindings button (same as above).
- New art for a empty portrait, used for example when you have an empty
item slot selected in the inventory.
Changes:
- Removed the first save statue from MysteriousDungeon, it didn't see
its use since you save the last thing you do before entering it and it
sort of got in the way.
- Added water at the end of the fall at the start of MysteriousDungeon.
- Removed the sign at the start of MysteriousDungeon, it didn't really
seem to serve much of a purpose.
- Heavily improved the edit key-bindings frame by turning it into a
JFrame that's inside of a JDialog module, thus that "window" will always
be above the game itself, it doesn't even let you alter the game for as
long as it's up. Thus you can no longer accidentally or otherwise open
more than one at any time, nor can you go back to playing the game with
the window still up.
- Portraits no longer uses the buff icon background, I went back to all
doodads, items and units using their own portrait again.