Version: 0.1.085
Changes:
- Items now make use of the use item keybinds meaning you can actually
change which key you for example drink a healing potion with.
- Items can now be moved around using the keyboard, use the interact
button to select an item you wish to move, then navigate to the slot you
wish to move it to, then press the interact button again. This can also
be used in order to swap the location of two items.
- Items now display their information when selected using the keyboard.

Bug fixes:
- Fixed two crashes that occured if you attempted to load LorasCavern
after talking to Liadrin.