Version: 0.1.086
Added:
- EmotionBubbles, they are meant to better display creatures emotions.
So far I only had the art for the exclamation one, which is displayed if
you try to loot more items than you can carry, for example.
- New item, BackPack. There is currently no way of obtaining it other
than "cheating" (pressing I). It uses the potion art because it doesn't
have its own yet. But it does work, by adding 5 new inventory slots when
used (and thus consumed). The new maximum number of inventory slots is
100 (10 * 10).

Changes:
- Added another key-bind that currently defaults to R and allows you to
toss items out of your inventory.
- If you try to pick up more items than you can currently carry you will
automatically toss the new items that go over the limit on the ground.
- Altered the way inventory slots are displayed a bit, it now supports
way more inventory slots in total, in preparation to allow the player to
have perhaps up to 100 inventory sltos (which is what I tested and it
worked well).
- Made some praparations to play "CannotCarry" sounds, the sound files
themselves have not yet been added.
- Reduced the number of inventory slots from the previous 25 to 5, I
will add ways of increasing this later (perhaps an item such as a
backpack, or a talent (if I ever implement those)).