Version: 0.1.097
Added:
- New sound for thunder, currently used whenever it starts raining for
extra effect. (It is a little bit loud on purpose, if you find it to be
too loud you can always reduce the volume for background sounds in the
options (yey options! ^^)).

Changes:
- The first map will now load automatically following the completion of
the tutorial.

Bug fixes:
- The player should now properly take damage in the dialogue with
Liadrin.
- There was apparently a bug that allowed the player to continuously
pick up items from the ground that they had already picked up, giving
the player access to an infinite amount of the item, this has been
fixed...
- Fixed a null-pointer error that prevented the player from going back
to the main menu from the tutorial.

Note:
- The tutorial is now completed, I will probably not make any further
additions or changes to it.