Version: 0.1.098
Added:
- Breath system:
When entering water there will now be a breath bar displayed showing
your current breath. It will automatically replenish if you leave the
water. If it reaches zero you will start losing health until you either
die or leave the water.
- Heat bar:
A new bar that shows your current heat. Later it will fade out if your
heat hasn't changed for a couple of seconds. I'm not sure if I will
replace the heat buffs with this completely or simply allow the player
to choose... I feel that the bar better shows your current value but the
buff gives a better feeling overall. I'll have to think about it for a
bit...
- Wet bar:
A new wet bar that tracks how wet you are. It should fade out when you
are completely dry.